Released landscape mode Released loading next pages Some refactor Preparations for BUY/RENT movie activity

+++ app/src/main/java/com/example/mymovies/MainActivity.java
@@ -1,22 +1,27 @@
 package com.example.mymovies;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.os.Bundle;
+import android.util.DisplayMetrics;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.CompoundButton;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
 
@@ -29,15 +34,22 @@ import com.example.mymovies.utils.NetworkUtils;
 import org.json.JSONObject;
 
 
+import java.net.URL;
 import java.util.List;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
 
     private Switch switchSort;
     private RecyclerView recyclerViewPosters;
     private MovieAdapter movieAdapter;
     private TextView textViewPopularity;
     private TextView textViewMostRated;
+    private static int LOADER_ID = 0;
+    private static int page = 1;
+    private static boolean isLoading = false;
+    private LoaderManager loaderManager;
+    private static int methodOfSort;
+    private ProgressBar progressBarLoading;
 
     private MainViewModel viewModel;
 
@@ -64,22 +76,32 @@ public class MainActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
+    private int getColumnCount() {
+        DisplayMetrics metrics = new DisplayMetrics();
+        getWindowManager().getDefaultDisplay().getMetrics(metrics);
+        int width = (int) (metrics.widthPixels / metrics.density);
+        return width/185 > 2 ? width / 185 : 2;
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        progressBarLoading = findViewById(R.id.progressBarLoading);
         viewModel = new ViewModelProvider(this).get(MainViewModel.class);
+        loaderManager = LoaderManager.getInstance(this);
         textViewPopularity = findViewById(R.id.textViewPopularity);
         textViewMostRated = findViewById(R.id.textViewMostRated);
         switchSort = findViewById(R.id.switchSort);
         recyclerViewPosters = findViewById(R.id.recyclerViewPosters);
-        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, 2));
+        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, getColumnCount()));
         movieAdapter = new MovieAdapter();
         switchSort.setChecked(true);
         recyclerViewPosters.setAdapter(movieAdapter);
         switchSort.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                page = 1;
                 setMethodOfSort(isChecked);
             }
         });
@@ -96,6 +118,9 @@ public class MainActivity extends AppCompatActivity {
         movieAdapter.setOnReachEndListener(new MovieAdapter.OnReachEndListener() {
             @Override
             public void onReachEnd() {
+                if(!isLoading){
+                    downloadData(methodOfSort, page);
+                }
 
             }
         });
@@ -103,7 +128,9 @@ public class MainActivity extends AppCompatActivity {
         moviesFromLiveData.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                movieAdapter.setMovies(movies);
+                if (page==1){
+                    movieAdapter.setMovies(movies);
+                }
             }
         });
     }
@@ -119,7 +146,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void setMethodOfSort(boolean isMostRated){
-        int methodOfSort;
         if (isMostRated) {
             methodOfSort = NetworkUtils.MOST_RATED;
             textViewMostRated.setTextColor(ContextCompat.getColor(this, R.color.green));
@@ -134,17 +160,52 @@ public class MainActivity extends AppCompatActivity {
             textViewPopularity.setTypeface(null, Typeface.BOLD);
             textViewMostRated.setTypeface(null, Typeface.NORMAL);
         }
-        downloadData(methodOfSort,1);
+        downloadData(methodOfSort,page);
     }
 
     private void downloadData(int methodOfSort, int page){
-        JSONObject jsonObject = NetworkUtils.getJSONFromNetwork(methodOfSort, 1);
-        List<Movie> movies = JSONUtils.getMoviesFromJSON(jsonObject);
+        URL url = NetworkUtils.buildURL(methodOfSort, page);
+        Bundle bundle = new Bundle();
+        bundle.putString("url", url.toString());
+        loaderManager.restartLoader(LOADER_ID, bundle, this);
+    }
+
+    @NonNull
+    @Override
+    public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle args) {
+        NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this, args);
+        jsonLoader.setOnStartLoadingListener(new NetworkUtils.JSONLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
+        return jsonLoader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
+        List<Movie> movies = JSONUtils.getMoviesFromJSON(data);
         if (movies!=null && !movies.isEmpty()){
-            viewModel.deleteAllMovies();
+            if (page == 1){
+                viewModel.deleteAllMovies();
+                movieAdapter.clear();
+            }
+
             for(Movie movie: movies){
                 viewModel.insertMovie(movie);
             }
+            movieAdapter.addMovies(movies);
+            page++;
         }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
+        loaderManager.destroyLoader(LOADER_ID);
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<JSONObject> loader) {
+
     }
 }
